+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Next release
 
 * Add paginated variant of `Users.lists`. The original method is deprecated.
-* Change to `https://auth.trakt.tv` for authentication requests as recommended by https://docs.trakt.tv/reference/getoauthauthorize.             
+* Change to `https://auth.trakt.tv` for authentication requests as recommended by https://docs.trakt.tv/reference/getoauthauthorize.
+* Change `TraktLink` to link to the new web app at `https://app.trakt.tv`. Deprecate no longer working external ID, season and episode links.
 
 ## 6.20.1 - 2026-06-18